移除受影响对象校验...

+++ codebase/netmarkets/jsp/ext/appo/change/request/removeAffectedLink.jsp
@@ -13,37 +13,68 @@
 <%@ page import="wt.part.WTPart" %>
 <%@ page import="wt.pom.Transaction" %>
 <%@ page import="wt.session.SessionHelper" %>
+<%@ page import="java.util.HashMap" %>
+<%@ page import="java.util.Map" %>
 <%@page pageEncoding="UTF-8" %>
 <%
     WTPrincipal principal = SessionHelper.manager.getPrincipal();
     SessionHelper.manager.setAdministrator();
-    Transaction tx = null;
-    try {
-        tx = new Transaction();
-        tx.start();
 
-        String oid = request.getParameter("oid");
-        String ecnVid = oid.replaceFirst("VR:wt.change2.WTChangeOrder2:", "");
-        String selectOid = request.getParameter("selectOid");
+    String oid = request.getParameter("oid");
+    String ecnVid = oid.replaceFirst("VR:wt.change2.WTChangeOrder2:", "");
+    String selectOid = request.getParameter("selectOid");
+    System.out.println("=====oid: " + oid);
+    System.out.println("=====ecnVid: " + ecnVid);
+    System.out.println("=====selectOid: " + selectOid);
 
-        System.out.println("=====oid: " + oid);
-        System.out.println("=====ecnVid: " + ecnVid);
-        System.out.println("=====selectOid: " + selectOid);
+    JSONArray removeVid = new JSONArray();
+    Map<Persistable, CorrelationObjectLink> resultMap = new HashMap<>();
+    StringBuffer messages = new StringBuffer("存在以下受影响对象不允许移除：\n");
 
-        JSONArray removeVid = new JSONArray();
-        if (StringUtils.isNotEmpty(ecnVid) && StringUtils.isNotEmpty(selectOid)) {
-            JSONArray selects = new JSONArray(selectOid);
-            System.out.println("=====selects: " + selects);
-            for (int i = 0; i < selects.length(); i++) {
-                Persistable persistable = ModifyUtils.getPersistable(selects.getString(i));
-                System.out.println("=====persistable: " + persistable);
-                if (persistable == null) continue;
+    //检查所选数据是否可以移除
+    if (StringUtils.isNotEmpty(ecnVid) && StringUtils.isNotEmpty(selectOid)) {
+        JSONArray selects = new JSONArray(selectOid);
+        System.out.println("=====selects: " + selects);
+        for (int i = 0; i < selects.length(); i++) {
+            String select = selects.getString(i);
+            System.out.println("=====select: " + select);
+            removeVid.put(select);
+
+            Persistable persistable = ModifyUtils.getPersistable(select);
+            System.out.println("=====persistable: " + persistable);
+            //未找到所选ID对应的对象，直接从受影响对象列表移除
+            if (persistable == null) continue;
+
+            String branchId = String.valueOf(PICoreHelper.service.getBranchId(persistable));
+            System.out.println("=====branchId: " + branchId);
+            CorrelationObjectLink link = ModifyHelper.service.queryCorrelationObjectLink(ecnVid, branchId, ModifyConstants.LINKTYPE_1);
+            System.out.println("=====link: " + link);
+            //未找到Link，直接从受影响对象列表移除
+            if (link == null) continue;
+
+            String routing = link.getRouting();
+            System.out.println("=====routing: " + routing);
+            if (StringUtils.isNotEmpty(routing) && !ModifyConstants.ROUTING_2.equals(routing)) {
+                messages.append(ModifyUtils.getNumber(persistable));
+                if (i < selects.length() - 1) messages.append("、");
+            }
+            resultMap.put(persistable, link);
+        }
+    }
+
+    if (messages.length() > 16) {
+        out.write(messages.toString());
+    } else {
+        Transaction tx = null;
+        try {
+            tx = new Transaction();
+            tx.start();
+
+            for (Map.Entry<Persistable, CorrelationObjectLink> entry : resultMap.entrySet()) {
+                Persistable persistable = entry.getKey();
+                CorrelationObjectLink link = entry.getValue();
 
                 //移除ECN与所选受影响对象的Link
-                String branchId = String.valueOf(PICoreHelper.service.getBranchId(persistable));
-                System.out.println("=====branchId: " + branchId);
-                CorrelationObjectLink link = ModifyHelper.service.queryCorrelationObjectLink(ecnVid, branchId, ModifyConstants.LINKTYPE_1);
-                System.out.println("=====link: " + link);
                 ModifyHelper.service.removeCorrelationObjectLink(link);
 
                 //所选对象为WTPart，移除关联的文档
@@ -55,24 +86,28 @@
                         if (object instanceof ObjectReference) {
                             object = ((ObjectReference) object).getObject();
                         }
-                        branchId = String.valueOf(PICoreHelper.service.getBranchId(object));
+                        String branchId = String.valueOf(PICoreHelper.service.getBranchId(object));
+                        if (StringUtils.isEmpty(branchId)) continue;
+
                         link = ModifyHelper.service.queryCorrelationObjectLink(ecnVid, branchId, ModifyConstants.LINKTYPE_1);
-                        if (link != null) removeVid.put(PICoreHelper.service.getVid(object));
+                        if (link == null) continue;
+
+                        removeVid.put(PICoreHelper.service.getVid(object));
                         ModifyHelper.service.removeCorrelationObjectLink(link);
                     }
                 }
-                removeVid.put(selects.getString(i));
+            }
+
+            tx.commit();
+            tx = null;
+        } finally {
+            if (tx != null) {
+                tx.rollback();
             }
         }
         System.out.println("=====removeVid: " + removeVid.toJSONString());
         out.write(removeVid.toJSONString());
-
-        tx.commit();
-        tx = null;
-    } finally {
-        if (tx != null) {
-            tx.rollback();
-        }
-        SessionHelper.manager.setPrincipal(principal.getName());
     }
+
+    SessionHelper.manager.setPrincipal(principal.getName());
 %>
